Using index template adapter to DRYing `IndicesAdapter`.
