+++ graylog-storage-elasticsearch7/src/test/java/org/graylog/storage/elasticsearch7/testing/AdaptersES7.java
@@ -18,6 +18,7 @@ package org.graylog.storage.elasticsearch7.testing;
 
 import com.codahale.metrics.MetricRegistry;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import org.graylog.storage.elasticsearch7.ComposableIndexTemplateAdapter;
 import org.graylog.storage.elasticsearch7.CountsAdapterES7;
 import org.graylog.storage.elasticsearch7.ElasticsearchClient;
 import org.graylog.storage.elasticsearch7.IndexFieldTypePollerAdapterES7;
@@ -69,7 +70,8 @@ public class AdaptersES7 implements Adapters {
                 new StatsApi(objectMapper, client),
                 new CatApi(objectMapper, client),
                 new ClusterStateApi(objectMapper, client),
-                objectMapper
+                objectMapper,
+                new ComposableIndexTemplateAdapter(client, objectMapper)
         );
     }
